playing card game
At the start of the game, there is one card on the table and all the players have five cards in their hand. A player can play a card from her/his hand if and only if it has the same rank or the same suit as the card on the table.
It is PrepBuddy's turn, there is a card on the table and five cards in her hand. Can you find out if she can play atleast one card?